refactor(spike): split oversized files under the 500-line/envelope cap
Per govern (code-governance) + CLAUDE.md file-size rule: split the 8 files exceeding the 300-500-line cap into probe + support modules, each now <400 lines and under the 24576-byte govern fleet envelope. Behavior preserved exactly (verified: py_compile clean, all probes offline exit 2/0 unchanged, decision.py INCONCLUSIVE, doctor result=skipped, no tracebacks).

decision.py->decision.py+decision_gates.py+decision_render.py; identity/partial_failure/fallback_inventory/snapshot/render/doctor/build each -> <probe>.py + <probe>_support.py.
